Split out stateful operator
