+++ operators.py
@@ -1,8 +1,9 @@
 ############### Operators ###############
+import imp
 import bpy, bgl, gpu
 from bpy.types import Operator
+
 from . import global_data, functions, class_defines, convertors
-from .keymaps import get_key_map_desc
 from .declarations import Operators, GizmoGroups, VisibilityTypes, WorkSpaceTools
 from . import preferences
 
@@ -544,1174 +545,18 @@ def get_hovered(context, *types):
             return entity
     return None
 
-
-def format_types(types):
-    entity_names = ", ".join([e.__name__ for e in types])
-    return "[" + entity_names + "]"
-
-
-def state_desc(name, desc, types):
-    type_desc = ""
-    if types:
-        type_desc = "Types: " + format_types(types)
-    return " ".join((name + ":", desc, type_desc))
-
-
-def stateful_op_desc(base, *state_descs):
-    states = ""
-    length = len(state_descs)
-    for i, state in enumerate(state_descs):
-        states += " - {}{}".format(
-            state, ("  \n" if i < length - 1 else "")
-        )
-    desc = "{}  \n  \nStates:  \n{}".format(base, states)
-    return desc
-
-
-numeric_events = (
-    "ZERO",
-    "ONE",
-    "TWO",
-    "THREE",
-    "FOUR",
-    "FIVE",
-    "SIX",
-    "SEVEN",
-    "EIGHT",
-    "NINE",
-    "PERIOD",
-    "NUMPAD_0",
-    "NUMPAD_1",
-    "NUMPAD_2",
-    "NUMPAD_3",
-    "NUMPAD_4",
-    "NUMPAD_5",
-    "NUMPAD_6",
-    "NUMPAD_7",
-    "NUMPAD_8",
-    "NUMPAD_9",
-    "NUMPAD_PERIOD",
-    "MINUS",
-    "NUMPAD_MINUS",
-)
-
-def get_evaluated_obj(context, object):
-    return object.evaluated_get(context.evaluated_depsgraph_get())
-
-def get_mesh_element(context, coords, vertex=False, edge=False, face=False, threshold=0.5, object=None):
-    from bpy_extras import view3d_utils
-
-    # get the ray from the viewport and mouse
-    region = context.region
-    rv3d = context.region_data
-    view_vector = view3d_utils.region_2d_to_vector_3d(region, rv3d, coords)
-    ray_origin = view3d_utils.region_2d_to_origin_3d(region, rv3d, coords)
-    depsgraph = context.view_layer.depsgraph
-    scene = context.scene
-    result, loc, _normal, face_index, ob, _matrix = scene.ray_cast(depsgraph, ray_origin, view_vector)
-
-    if object:
-        # Alternatively do a object raycast if we know the object already
-        tree = functions.bvhtree_from_object(ob)
-        loc, _normal, face_index, _distance = tree.ray_cast(ray_origin, view_vector)
-        result = loc != None
-        ob = object
-
-    if not result:
-        return None, None, None
-
-    obj_eval = get_evaluated_obj(context, ob)
-
-    closest_type = ""
-    closest_dist = None
-
-    loc = obj_eval.matrix_world.inverted() @ loc
-    me = obj_eval.data
-    polygon = me.polygons[face_index]
-
-    def get_closest(deltas):
-        index_min = min(range(len(deltas)), key=deltas.__getitem__)
-        if deltas[index_min] > threshold:
-            return None, None
-        return index_min, deltas[index_min]
-
-    def is_closer(distance, min_distance):
-        if min_distance is None:
-            return True
-        if distance < min_distance:
-            return True
-        return False
-
-    if vertex:
-        i, dist = get_closest([(me.vertices[i].co - loc).length for i in polygon.vertices])
-        if i is not None:
-            closest_type = "VERTEX"
-            closest_index = polygon.vertices[i]
-            closest_dist = dist
-
-    if edge:
-        face_edge_map = {ek: me.edges[i] for i, ek in enumerate(me.edge_keys)}
-        i, dist = get_closest(
-            [(((me.vertices[start].co + me.vertices[end].co)/2)-loc).length for start, end in polygon.edge_keys]
-            )
-        if i is not None and is_closer(dist, closest_dist):
-            closest_type = "EDGE"
-            closest_index = face_edge_map[polygon.edge_keys[i]].index
-            closest_dist = dist
-
-    if face:
-        # Check if face midpoint is closest
-        if is_closer((polygon.center - loc).length, closest_dist):
-            closest_type = "FACE"
-            closest_index = face_index
-
-    if closest_type:
-        return ob, closest_type, closest_index
-    return ob, bpy.types.Object, None
-
-def to_list(val):
-    if val == None:
-        return []
-    if type(val) in (list, tuple):
-        return list(val)
-    return [val,]
-
-def _get_pointer_get_set(index):
-    @property
-    def func(self):
-        return self.get_state_pointer(index=index)
-
-    @func.setter
-    def setter(self, value):
-        self.set_state_pointer(value, index=index)
-    return func, setter
-
-mesh_element_types = bpy.types.MeshVertex, bpy.types.MeshEdge, bpy.types.MeshPolygon
-
-class StatefulOperator:
-    state_index: IntProperty(options={"HIDDEN", "SKIP_SAVE"})
-    wait_for_input: BoolProperty(options={"HIDDEN", "SKIP_SAVE"}, default=True)
-    continuous_draw: BoolProperty(name="Continuous Draw", default=False)
-
-    executed = False
-    _state_data = {}
-    _last_coords = Vector((0, 0))
-    _numeric_input = {}
-    _undo = False
-
-    @classmethod
-    def _has_global_object(cls):
-        states = cls.get_states_definition()
-        return any([s.pointer == "object" for s in states])
-
-    def _get_global_object_index(cls):
-        states = cls.get_states_definition()
-        object_in_list = [s.pointer == "object" for s in states]
-        if not any(object_in_list):
-            return None
-        return object_in_list.index(True)
-
-    @classmethod
-    def register_properties(cls):
-        states = cls.get_states_definition()
-        annotations = cls.__annotations__.copy()
-
-        # Have some specific logic: pointer name "object" is used as global object
-        # otherwise define ob_name for each element
-        has_global_object = cls._has_global_object()
-
-        for i, s in enumerate(states):
-            pointer_name = s.pointer
-            types = s.types
-
-            if not pointer_name:
-                continue
-
-            if pointer_name in annotations.keys():
-                # Skip pointers that have a property defined
-                # Note: pointer might not need implicit props, thus no need for getter/setter
-                return
-
-            if hasattr(cls, pointer_name):
-                # This can happen when the addon is re-enabled in the same session
-                continue
-
-            get, set = _get_pointer_get_set(i)
-            setattr(cls, pointer_name, get)
-            # Note: keep state pointers read-only, only set with set_state_pointer()
-
-        for a in annotations.keys():
-            if hasattr(cls, a):
-                raise NameError("Cannot register implicit pointer properties, class {} already has attribute of name {}".format(cls, a))
-
-
-    def state_property(self, state_index):
-        return None
-
-    def get_state_pointer(self, index=None, implicit=False):
-        # Creates pointer value from it's implicitly stored props
-        if index is None:
-            index = self.state_index
-
-        state = self.get_states_definition()[index]
-        pointer_name = state.pointer
-        data = self._state_data.get(index, {})
-        if not "type" in data.keys():
-            return None
-
-        pointer_type = data["type"]
-        if not pointer_type:
-            return None
-
-        if pointer_type in (bpy.types.Object, *mesh_element_types):
-            if self._has_global_object():
-                global_ob_index = self._get_global_object_index()
-                obj_name = self._state_data[global_ob_index]["object_name"]
-            else:
-                obj_name = data["object_name"]
-            obj = get_evaluated_obj(bpy.context, bpy.data.objects[obj_name])
-
-        if pointer_type in mesh_element_types:
-            index = data["mesh_index"]
-
-        if pointer_type == bpy.types.Object:
-            if implicit:
-                return obj_name
-            return obj
-
-        elif pointer_type == bpy.types.MeshVertex:
-            if implicit:
-                return obj_name, index
-            return obj.data.vertices[index]
-
-        elif pointer_type == bpy.types.MeshEdge:
-            if implicit:
-                return obj_name, index
-            return obj.data.edges[index]
-
-        elif pointer_type == bpy.types.MeshPolygon:
-            if implicit:
-                return obj_name, index
-            return obj.data.polygons[index]
-
-
-    def set_state_pointer(self, values, index=None, implicit=False):
-        # handles type specific setters
-        if index is None:
-            index = self.state_index
-
-        state = self.get_states_definition()[index]
-        pointer_name = state.pointer
-        data = self._state_data.get(index, {})
-
-        pointer_type = data["type"]
-
-        def get_value(index):
-            if values == None:
-                return None
-            return values[index]
-
-        if pointer_type == bpy.types.Object:
-            if implicit:
-                val = get_value(0)
-            else:
-                val = get_value(0).name
-            data["object_name"] = val
-            return True
-
-        elif pointer_type in mesh_element_types:
-            obj_name = get_value(0) if implicit else get_value(0).name
-            if self._has_global_object():
-                self._state_data[self._get_global_object_index()]["object_name"] = obj_name
-            else:
-                data["object_name"] = obj_name
-
-            data["mesh_index"] = get_value(1) if implicit else get_value(1).index
-            return True
-
-
-    def pick_element(self, context, coords):
-        # return a list of implicit prop values if pointer need implicit props
-        state = self.state
-        data = self.state_data
-
-        types = {
-            "vertex": (bpy.types.MeshVertex in state.types),
-            "edge": (bpy.types.MeshEdge in state.types),
-            "face": (bpy.types.MeshPolygon in state.types),
-        }
-
-        do_object = bpy.types.Object in state.types
-        do_mesh_elem = any(types.values())
-
-        if not do_object and not do_mesh_elem:
-            return
-
-        global_ob = None
-        if self._has_global_object():
-            global_ob_name = self._state_data[self._get_global_object_index()].get("object_name")
-            if global_ob_name:
-                global_ob = bpy.data.objects[global_ob_name]
-
-        ob, type, index = get_mesh_element(context, coords, **types, object=global_ob)
-
-        if not ob:
-            return None
-
-        if bpy.types.Object in state.types:
-            data["type"] = bpy.types.Object
-            return ob.name
-
-        # maybe have type as return value
-        data["type"] = {
-            "VERTEX": bpy.types.MeshVertex,
-            "EDGE": bpy.types.MeshEdge,
-            "FACE": bpy.types.MeshPolygon,
-        }[type]
-
-        return ob.name, index
-
-
-    def get_property(self, index=None):
-        if index == None:
-            index = self.state_index
-        state = self.get_states()[index]
-
-        if state.property == None:
-            return None
-
-        if callable(state.property):
-            props = state.property(self, index)
-        elif state.property:
-            if callable(getattr(self, state.property)):
-                props = getattr(self, state.property)(index)
-            else:
-                props = state.property
-        elif hasattr(self, "state_property") and callable(self.state_property):
-            props = self.state_property(index)
-        else:
-            return None
-
-        retval = to_list(props)
-        return retval
-
-
-    @classmethod
-    def get_states_definition(cls):
-        if callable(cls.states):
-            return cls.states()
-        return cls.states
-
-    def get_states(self):
-        if callable(self.states):
-            return self.states(operator=self)
-        return self.states
-
-    @property
-    def state(self):
-        return self.get_states()[self.state_index]
-
-    def _index_from_state(self, state):
-        return [e.name for e in self.get_states()].index(state)
-
-    @state.setter
-    def state(self, state):
-        self.state_index = self._index_from_state(state)
-
-    def set_state(self, context, index):
-        self.state_index = index
-        self.init_numeric(False)
-        self.set_status_text(context)
-
-    def next_state(self, context):
-        self._undo = False
-        i = self.state_index
-        if (i + 1) >= len(self.get_states()):
-            return False
-        self.set_state(context, i + 1)
-        return True
-
-    def set_status_text(self, context):
-        # Setup state
-        state = self.state
-        desc = (
-            state.description(self, state)
-            if callable(state.description)
-            else state.description
-        )
-
-        msg = state_desc(state.name, desc, state.types)
-        if self.state_data.get("is_numeric_edit", False):
-            index = self._substate_index
-            prop = self._stateprop
-            type = prop.type
-            array_length = prop.array_length if prop.array_length else 1
-            if type == "FLOAT":
-                input = [0.0] * array_length
-                for key, val in self._numeric_input.items():
-                    input[key] = val
-
-                input[index] = "*" + str(input[index])
-                input = str(input).replace('"', "").replace("'", "")
-            elif type == "INT":
-                input = self.numeric_input
-
-            msg += "    {}: {}".format(prop.subtype, input)
-
-        context.workspace.status_text_set(msg)
-
-    def check_numeric(self):
-        """Check if the state supports numeric edit"""
-
-        # TODO: Allow to define custom logic
-
-        state = self.state
-        props = self.get_property()
-
-        # Disable for multi props
-        if not props or len(props) > 1:
-            return False
-
-        prop_name = props[0]
-        if not prop_name:
-            return False
-
-        prop = self.properties.rna_type.properties[prop_name]
-        if not prop.type in ("INT", "FLOAT"):
-            return False
-        return True
-
-    def init_numeric(self, is_numeric):
-        self._numeric_input = {}
-        self._substate_index = 0
-
-        ok = False
-        if is_numeric:
-            ok = self.check_numeric()
-            # TODO: not when iterating substates
-            self.state_data["is_numeric_edit"] = is_numeric and ok
-
-        self.init_substate()
-        return ok
-
-    def init_substate(self):
-        props = self.get_property()
-        if props and props[0]:
-            prop_name = props[0]
-            prop = self.properties.rna_type.properties[prop_name]
-            self._substate_count = prop.array_length
-            self._stateprop = prop
-        else:
-            self._substate_count = None
-            self._stateprop = None
-
-    def iterate_substate(self):
-        i = self._substate_index
-        if i + 1 >= self._substate_count:
-            i = 0
-        else:
-            i += 1
-        self._substate_index = i
-
-    @property
-    def numeric_input(self):
-        return self._numeric_input.get(self._substate_index, "")
-
-    @numeric_input.setter
-    def numeric_input(self, value):
-        self._numeric_input[self._substate_index] = value
-
-    def check_event(self, event):
-        state = self.state
-
-        if event.type == "LEFT_SHIFT":
-            bpy.context.scene.sketcher.selectable_constraints = event.value == "RELEASE"
-
-        if (
-            event.type in ("LEFTMOUSE", "RET", "NUMPAD_ENTER")
-            and event.value == "PRESS"
-        ):
-            return True
-        if self.state_index == 0 and not self.wait_for_input:
-            # Trigger the first state
-            return not self.state_data.get("is_numeric_edit", False)
-        if state.no_event:
-            return True
-        return False
-
-    @staticmethod
-    def is_numeric_input(event):
-        return event.type in (*numeric_events, "BACK_SPACE")
-
-    @staticmethod
-    def is_unit_input(event):
-        return event.type in (
-            "M",
-            "K",
-            "D",
-            "C",
-            "U",
-            "A",
-            "H",
-            "I",
-            "L",
-            "N",
-            "F",
-            "T",
-            "Y",
-            "U",
-            "R",
-            "E",
-            "G",
-        )
-
-    @staticmethod
-    def get_unit_value(event):
-        type = event.type
-        return type.lower()
-
-    @staticmethod
-    def get_value_from_event(event):
-        type = event.type
-        if type in ("ZERO", "NUMPAD_0"):
-            return "0"
-        if type in ("ONE", "NUMPAD_1"):
-            return "1"
-        if type in ("TWO", "NUMPAD_2"):
-            return "2"
-        if type in ("THREE", "NUMPAD_3"):
-            return "3"
-        if type in ("FOUR", "NUMPAD_4"):
-            return "4"
-        if type in ("FIVE", "NUMPAD_5"):
-            return "5"
-        if type in ("SIX", "NUMPAD_6"):
-            return "6"
-        if type in ("SEVEN", "NUMPAD_7"):
-            return "7"
-        if type in ("EIGHT", "NUMPAD_8"):
-            return "8"
-        if type in ("NINE", "NUMPAD_9"):
-            return "9"
-        if type in ("PERIOD", "NUMPAD_PERIOD"):
-            return "."
-
-    def evaluate_numeric_event(self, event):
-        type = event.type
-        if type == "BACK_SPACE":
-            input = self.numeric_input
-            if len(input):
-                self.numeric_input = input[:-1]
-        elif type in ("MINUS", "NUMPAD_MINUS"):
-            input = self.numeric_input
-            if input.startswith("-"):
-                input = input[1:]
-            else:
-                input = "-" + input
-            self.numeric_input = input
-        elif self.is_unit_input(event):
-            self.numeric_input += self.get_unit_value(event)
-        else:
-            self.numeric_input += self.get_value_from_event(event)
-
-    def is_in_previous_states(self, entity):
-        i = self.state_index - 1
-        while True:
-            if i < 0:
-                break
-            state = self.get_states()[i]
-            if state.pointer and entity == getattr(self, state.pointer):
-                return True
-            i -= 1
-        return False
-
-    def gather_selection(self, context):
-        # Return list filled with all selected verts/edges/faces/objects
-        selected = []
-        states = self.get_states()
-        types = [s.types for s in states]
-        # Note: Where to take mesh elements from? Editmode data is only written
-        # when left probably making it impossible to use selected elements in realtime.
-        if any([t == bpy.types.Object for t in types]):
-            selected.extend(context.selected_objects)
-
-        return selected
-
-    # Gets called for every state
-    def parse_selection(self, context, selected, index=None):
-        # Look for a valid element in selection
-        # should go through objects, vertices, entities depending on state.types
-
-        result = None
-        if not index:
-            index = self.state_index
-        state = self.get_states_definition()[index]
-        data = self.get_state_data(index)
-
-        if state.pointer:
-            # TODO: Discard if too many entities are selected?
-            types = state.types
-            for i, e in enumerate(selected):
-                if type(e) in types:
-                    result = selected.pop(i)
-                    break
-
-        if result:
-            data["type"] = type(result)
-            self.set_state_pointer(to_list(result), index=index)
-            self.state_data["is_existing_entity"] = True
-            return True
-
-
-    def prefill_state_props(self, context):
-        selected = self.gather_selection(context)
-        states = self.get_states_definition()
-
-        # Iterate states and try to prefill state props
-        while True:
-            index = self.state_index
-            result = None
-            state = self.state
-            data = self.get_state_data(index)
-            coords = None
-
-            if not state.allow_prefill:
-                break
-
-            func = self.get_func(state, "parse_selection")
-            result = func(context, selected, index=index)
-
-            if result:
-                if not self.next_state(context):
-                    return {"FINISHED"}
-                continue
-            break
-        return {"RUNNING_MODAL"}
-
-    @property
-    def state_data(self):
-        return self._state_data.setdefault(self.state_index, {})
-
-    def get_state_data(self, index):
-        if not self._state_data.get(index):
-            self._state_data[index] = {}
-        return self._state_data[index]
-
-    def get_func(self, state, name):
-        # fallback to operator method if function isn't specified by state
-        func = getattr(state, name, None)
-
-        if func:
-            if isinstance(func, str):
-                # callback can be specified by function name
-                return getattr(self, func)
-            return func
-
-        if hasattr(self, name):
-            return getattr(self, name)
-        return None
-
-    def invoke(self, context, event):
-        self._state_data.clear()
-        if hasattr(self, "init"):
-            self.init(context, event)
-
-        retval = {"RUNNING_MODAL"}
-
-        go_modal = True
-        if self.is_numeric_input(event):
-            if self.init_numeric(True):
-                self.evaluate_numeric_event(event)
-                retval = {"RUNNING_MODAL"}
-                self.evaluate_state(context, event, False)
-
-        # NOTE: This allows to start the operator but waits for action (LMB event).
-        # Try to fill states based on selection only when this is True since it doesn't
-        # make senese to respect selection when the user interactivley starts the operator.
-        elif self.wait_for_input:
-            retval = self.prefill_state_props(context)
-            if retval == {"FINISHED"}:
-                go_modal = False
-
-            # NOTE: It might make sense to cancel Operator if no prop could be filled
-            # Otherwise it might not be obvious that an operator is running
-            # if self.state_index == 0:
-            #     return self._end(context, False)
-
-            if not self.executed and self.check_props():
-                self.run_op(context)
-                self.executed = True
-            context.area.tag_redraw()  # doesn't seem to work...
-
-        self.set_status_text(context)
-
-        if go_modal:
-            context.window.cursor_modal_set("CROSSHAIR")
-            context.window_manager.modal_handler_add(self)
-            return retval
-
-        succeede = retval == {"FINISHED"}
-        if succeede:
-            # NOTE: It seems like there's no undo step pushed if an operator finishes from invoke
-            # could push an undo_step here however this causes duplicated constraints after redo,
-            # disable for now
-            # bpy.ops.ed.undo_push()
-            pass
-        return self._end(context, succeede)
-
-    def run_op(self, context):
-        if not hasattr(self, "main"):
-            raise NotImplementedError(
-                "StatefulOperators need to have a main method defined!"
-            )
-        retval = self.main(context)
-        self.executed = True
-        return retval
-
-    # Creates non-persistent data
-    def redo_states(self, context):
-        for i, state in enumerate(self.get_states()):
-            if i > self.state_index:
-                # TODO: don't depend on active state, idealy it's possible to go back
-                break
-            if state.pointer:
-                data = self._state_data.get(i, {})
-                is_existing_entity = data["is_existing_entity"]
-
-                props = self.get_property(index=i)
-                if props and not is_existing_entity:
-                    create = self.get_func(state, "create_element")
-
-                    ret_values = create(context, [getattr(self, p) for p in props], state, data)
-                    values = to_list(ret_values)
-                    self.set_state_pointer(values, index=i, implicit=True)
-
-    def execute(self, context):
-        self.redo_states(context)
-        ok = self.main(context)
-        return self._end(context, ok)
-        # maybe allow to be modal again?
-
-    def get_numeric_value(self, context, coords):
-        state = self.state
-        prop_name = self.get_property()[0]
-        prop = self.properties.rna_type.properties[prop_name]
-
-        def parse_input(prop, input):
-            units = context.scene.unit_settings
-            unit = prop.unit
-            type = prop.type
-            value = None
-
-            if input == "-":
-                pass
-            elif unit != "NONE":
-                try:
-                    value = bpy.utils.units.to_value(units.system, unit, input)
-                except ValueError:
-                    return prop.default
-                if type == "INT":
-                    value = int(value)
-            elif type == "FLOAT":
-                value = float(input)
-            elif type == "INT":
-                value = int(input)
-
-            if value == None:
-                return prop.default
-            return value
-
-        size = max(1, self._substate_count)
-
-        def to_iterable(item):
-            if hasattr(item, "__iter__") or hasattr(item, "__getitem__"):
-                return list(item)
-            return [item, ]
-
-        # TODO: Don't evaluate if not needed
-        interactive_val = self._get_state_values(context, state, coords)
-        if interactive_val == None:
-            interactive_val = [None] * size
-        else:
-            interactive_val = to_iterable(interactive_val)
-
-        storage = [None] * size
-        result = [None] * size
-
-        for sub_index in range(size):
-            num = None
-
-            input = self._numeric_input.get(sub_index)
-            if input:
-                num = parse_input(prop, input)
-                result[sub_index] = num
-                storage[sub_index] = num
-            elif interactive_val[sub_index] != None:
-                result[sub_index] = interactive_val[sub_index]
-            else:
-                result[sub_index] = prop.default
-
-        self.state_data["numeric_input"] = storage
-
-        if not self._substate_count:
-            return result[0]
-        return result
-
-    def _handle_pass_through(self, context, event):
-        # Only pass through navigation events
-        if event.type in {'MIDDLEMOUSE', 'WHEELUPMOUSE', 'WHEELDOWNMOUSE', "MOUSEMOVE"}:
-            return {"PASS_THROUGH"}
-        return {"RUNNING_MODAL"}
-
-    def modal(self, context, event):
-        state = self.state
-        event_triggered = self.check_event(event)
-        coords = Vector((event.mouse_region_x, event.mouse_region_y))
-
-        is_numeric_edit = self.state_data.get("is_numeric_edit", False)
-        is_numeric_event = event.value == "PRESS" and self.is_numeric_input(event)
-
-        if is_numeric_edit:
-            if self.is_unit_input(event) and event.value == "PRESS":
-                is_numeric_event = True
-            elif event.type == "TAB" and event.value == "PRESS":
-                self.iterate_substate()
-                self.set_status_text(context)
-        elif is_numeric_event:
-            # Initialize
-            is_numeric_edit = self.init_numeric(True)
-
-        if event.type in {"RIGHTMOUSE", "ESC"}:
-            return self._end(context, False)
-
-        # HACK: when calling ops.ed.undo() inside an operator a mousemove event
-        # is getting triggered. manually check if there's a mousemove...
-        mousemove_threshold = 0.1
-        is_mousemove = (coords - self._last_coords).length > mousemove_threshold
-        self._last_coords = coords
-
-        if not event_triggered:
-            if is_numeric_event:
-                pass
-            elif is_mousemove and is_numeric_edit:
-                event_triggered = False
-                pass
-            elif not state.interactive:
-                return self._handle_pass_through(context, event)
-            elif not is_mousemove:
-                return self._handle_pass_through(context, event)
-
-        # TODO: Disable numeric input when no state.property
-        if is_numeric_event:
-            self.evaluate_numeric_event(event)
-            self.set_status_text(context)
-
-        return self.evaluate_state(context, event, event_triggered)
-
-    def _get_state_values(self, context, state, coords):
-        # Get values of state_func, can be none
-        position_cb = self.get_func(state, "state_func")
-        if not position_cb:
-            return None
-        pos_val = position_cb(context, coords)
-        if pos_val == None:
-            return None
-        return pos_val
-
-    def evaluate_state(self, context, event, triggered):
-        state = self.state
-        data = self.state_data
-        is_numeric = self.state_data.get("is_numeric_edit", False)
-        coords = Vector((event.mouse_region_x, event.mouse_region_y))
-
-        # Pick hovered element
-        hovered = None
-        is_picked = False
-        if not is_numeric and state.pointer:
-            pick = self.get_func(state, "pick_element")
-            pick_retval = pick(context, coords)
-
-            if pick_retval != None:
-                is_picked = True
-                pointer_values = to_list(pick_retval)
-
-        # Set state property
-        ok = False
-        values = []
-        use_create = state.use_create
-        if use_create and not is_picked:
-            if is_numeric:
-                # numeric edit is supported for one property only
-                values = [self.get_numeric_value(context, coords), ]
-            elif not is_picked:
-                values = to_list(self._get_state_values(context, state, coords))
-
-            if values:
-                props = self.get_property()
-                if props:
-                    for i, v in enumerate(values):
-                        setattr(self, props[i], v)
-                    self._undo = True
-                    ok = not state.pointer
-
-        # Set state pointer
-        pointer = None
-        if state.pointer:
-            if is_picked:
-                pointer = pointer_values
-                self.state_data["is_existing_entity"] = True
-            elif values:
-                # Let pointer be filled from redo_states
-                self.state_data["is_existing_entity"] = False
-                ok = True
-
-            if pointer:
-                self.set_state_pointer(pointer, implicit=True)
-                ok = True
-
-        if self._undo:
-            bpy.ops.ed.undo_push(message="Redo: " + self.bl_label)
-            bpy.ops.ed.undo()
-            global_data.ignore_list.clear()
-            self.redo_states(context)
-            self._undo = False
-
-        if self.check_props():
-            self.run_op(context)
-            self._undo = True
-
-        # Iterate state
-        if triggered and ok:
-            if not self.next_state(context):
-                if self.check_continuous_draw():
-                    self.do_continuous_draw(context)
-                else:
-                    return self._end(context, True)
-
-            if is_numeric:
-                # NOTE: Run next state already once even if there's no mousemove yet,
-                # This is needed in order for the geometry to update
-                self.evaluate_state(context, event, False)
-        context.area.tag_redraw()
-
-
-        if triggered and not ok:
-            # Event was triggered on non-valid selection, cancel operator to avoid confusion
-            return self._end(context, False)
-
-        if triggered or is_numeric:
-            return {"RUNNING_MODAL"}
-        return self._handle_pass_through(context, event)
-
-    def check_continuous_draw(self):
-        if self.continuous_draw:
-            if not hasattr(self, "continue_draw") or self.continue_draw():
-                return True
-        return False
-
-    def _reset_op(self):
-        self.executed = False
-        for i, s in enumerate(self.get_states()):
-            if not s.pointer:
-                continue
-            self.set_state_pointer(None, index=i)
-        self._state_data.clear()
-
-    def do_continuous_draw(self, context):
-        # end operator
-        self._end(context, True)
-        bpy.ops.ed.undo_push(message=self.bl_label)
-
-        # save last prop
-        last_pointer = None
-        for i, s in reversed(list(enumerate(self.get_states()))):
-            if not s.pointer:
-                continue
-            last_index = i
-            last_pointer = getattr(self, s.pointer)
-            break
-
-        values = to_list(self.get_state_pointer(index=last_index, implicit=True))
-
-        # reset operator
-        self._reset_op()
-
-        data = {}
-        self._state_data[0] = data
-        data["is_existing_entity"] = True
-        data["type"] = type(last_pointer)
-
-        # set first pointer
-        self.set_state_pointer(values, index=0, implicit=True)
-        self.set_state(context, 1)
-
-    def _end(self, context, succeede):
-        context.window.cursor_modal_restore()
-        if hasattr(self, "fini"):
-            self.fini(context, succeede)
-        global_data.ignore_list.clear()
-
-        context.workspace.status_text_set(None)
-        if succeede:
-            return {"FINISHED"}
-        else:
-            bpy.ops.ed.undo_push(message="Cancelled: " + self.bl_label)
-            bpy.ops.ed.undo()
-            return {"CANCELLED"}
-
-    def check_props(self):
-        for i, state in enumerate(self.get_states()):
-            props = self.get_property(index=i)
-            if state.pointer:
-                if not bool(self.get_state_pointer(index=i)):
-                    return False
-
-            elif props:
-                for p in props:
-                    if not self.properties.is_property_set(p):
-                        return False
-        return True
-
-    def draw(self, context):
-        layout = self.layout
-
-        for i, state in enumerate(self.get_states()):
-            if i != 0:
-                layout.separator()
-
-            layout.label(text=state.name)
-
-            state_data = self._state_data.get(i, {})
-            is_existing = state_data.get("is_existing_entity", False)
-            props = self.get_property(index=i)
-
-            if state.pointer and is_existing:
-                layout.label(text=str(getattr(self, state.pointer)))
-            elif props:
-                for p in props:
-                    layout.prop(self, p, text="")
-
-        if hasattr(self, "draw_settings"):
-            self.draw_settings(context)
-
-    @classmethod
-    def description(cls, context, properties):
-        states = [state_desc(s.name, s.description, s.types) for s in cls.get_states_definition()]
-        descs = []
-        hint = get_key_map_desc(cls.bl_idname)
-        if hint:
-            descs.append(hint)
-        if cls.__doc__:
-            descs.append(cls.__doc__)
-        return stateful_op_desc(" ".join(descs), *states)
-
-
-# StatefulOperator Doc
-
-# Operator Methods
-# main(self, context) -> succeede(bool)
-#   function which creates the actual result of the operator,
-#   e.g. the main function of an add_line operator creates the line
-
-# register_properties(cls) -> None
-#   can be used to store implicit pointer properties and pointer fallback properties
-
-# get_state_pointer(self, index=None) -> result(ANY)
-#   method for pointer access, either of active state or by state index
-
-# set_state_pointer(self, value, index=None) -> succeede(bool)
-#   method to set state pointer, either of active state or by state index
-
-# check_props(self) -> succeede(bool)
-#   additional poll function to check if all necessary operator properties
-#   are set and the main function can be called
-
-# init(self, context, event) -> None
-
-# fini(self, context, succeede) -> None
-
-# check_pointer(self, prop_name) -> is_set(bool)
-#   check if a state pointer is set
-
-# gather_selection(self, context) -> selected(list(ANY))
-#   gather the currently selected elements that are later used to fill state pointers with
-
-# State Definition
-# state_func(self, context, coords) property_value(ANY)
-#   method to get the value for the state property from mouse coordinates
-
-# pick_element(self, context, coords) -> element or it's implicit props
-#   method to pick a matching element from mouse coordinates, either return the
-#   element or it's implicit prop values, has to set the type of the picked element
-
-# create_element(self, context, value, state, state_data) -> element or it's implicit props
-#   method to create state element when no existing element gets picked,
-#   has to set the type of the created element
-
-
-from collections import namedtuple
-
-OperatorState = namedtuple(
-    "OperatorState",
-    (
-        "name",  # The name to display in the interface
-        "description",  # Text to be displayed in statusbar
-        # Operator property this state acts upon
-        # Can also be a list of property names or a callback that returns
-        # a set of properties dynamically. When not explicitly set to None the
-        # operators state_property function will be called.
-        "property",
-        # Optional: A state can reference an element, pointer attribute set the name of property function
-        # if set this will be passed to main func,
-        # state_func should fill main property and create_element should fill this property
-        # maybe this could just store it in a normal attr, should work as long as the same operator instance is used, test!
-        "pointer",
-        "types",  # Types the pointer property can accept
-        "no_event",  # Trigger state without an event
-        "interactive",  # Always evaluate state and confirm by user input
-        "use_create", # Enables or Disables creation of the element
-        "state_func",  # Function to get the state property value from mouse coordinates
-        "allow_prefill",  # Define if state should be filled from selected entities when invoked
-        "parse_selection",  # Prefill Function which chooses entity to use for this stat
-        "pick_element",
-        "create_element",
-        # TODO: Implement!
-        "use_interactive_placemenet",  # Create new state element based on mouse coordinates
-        "check_pointer",
-    ),
-)
-del namedtuple
-
-
-def state_from_args(name, **kwargs):
-    """
-    Use so each state can avoid defining all members of the named tuple.
-    """
-    kw = {
-        "name": name,
-        "description": "",
-        "property": "",
-        "pointer": None,
-        "types": (),
-        "no_event": False,
-        "interactive": False,
-        "use_create" : True,
-        "state_func": None,
-        "allow_prefill": True,
-        "parse_selection": None,
-        "pick_element": None,
-        "create_element": None,
-        "use_interactive_placemenet": True,
-        "check_pointer": None,
-    }
-    kw.update(kwargs)
-    return OperatorState(**kw)
-
-
-from bpy_extras.view3d_utils import region_2d_to_location_3d, region_2d_to_vector_3d
-
+from .stateful_operator.integration import StatefulOperator
+from .stateful_operator.state import state_from_args
 
 class GenericEntityOp(StatefulOperator):
+    """Extend StatefulOperator with addon specific types"""
+
+    def check_event(self, event):
+        # Hardcode shift event to toggle constraint selectability
+        if event.type == "LEFT_SHIFT":
+            bpy.context.scene.sketcher.selectable_constraints = event.value == "RELEASE"
+        return super().check_event(event)
+
     def pick_element(self, context, coords):
         retval = super().pick_element(context, coords)
         if retval != None:
@@ -2888,48 +1733,6 @@ class View3D_OT_slvs_test(Operator, GenericEntityOp):
         return False
 
 
-class View3D_OT_invoke_tool(Operator):
-    bl_idname = Operators.InvokeTool
-    bl_label = "Invoke Tool"
-
-    tool_name: StringProperty(name="Tool ID")
-
-    # TODO: get the operator from tool attribute (tool.bl_operator)?
-    operator: StringProperty(name="Operator ID")
-
-    def execute(self, context):
-        bpy.ops.wm.tool_set_by_id(name=self.tool_name)
-
-        # get the tool operator props
-        tool = context.workspace.tools.from_space_view3d_mode(context.mode)
-        props = tool.operator_properties(self.operator)
-
-        options = {}
-        for p in props.rna_type.properties.keys():
-            if p in ("bl_rna", "rna_type", "state_index"):
-                continue
-            if p.startswith("_"):
-                continue
-
-            default = props.rna_type.properties[p].default
-            value = getattr(props, p)
-
-            # NOTE: Setting all values might mess around with operators that check
-            # if properties are set. Checking is_property_set here doesn't seem to work.
-            # manually check if value is the same as the default however that also
-            # seems to fail for vectorproperties and maybe others but prevents
-            # the problems from caused by pointer set-state checking...
-            if value != default:
-                options[p] = value
-
-        options["wait_for_input"] = True
-
-        op_name = self.operator.split(".", 1)
-        op = getattr(getattr(bpy.ops, op_name[0]), op_name[1])
-        if op.poll():
-            op("INVOKE_DEFAULT", **options)
-        return {"FINISHED"}
-
 SMOOTHVIEW_FACTOR = 0
 def align_view(rv3d, mat_start, mat_end):
 
@@ -3760,6 +2563,8 @@ constraint_operators = (
     VIEW3D_OT_slvs_add_midpoint,
     VIEW3D_OT_slvs_add_ratio,
 )
+
+from .stateful_operator.invoke_op import View3D_OT_invoke_tool
 
 classes = (
     View3D_OT_slvs_register_draw_cb,
